+++ pyhbasecli/HBase.py
@@ -527,7 +527,7 @@ class MutationSession:
             raise ResubmitException("can not add mutation to a submited session")
         _columns = None
         if columns:
-            _columns = [TColumn(family=c.split(":")[0], qualifier=c.split(":")[1]) if len(c.split(":")) == 2 else TColumn(family=c) for c in columns]
+            _columns = [TColumn(family=c.split(":")[0].encode(), qualifier=c.split(":")[1].encode()) if len(c.split(":")) == 2 else TColumn(family=c.encode()) for c in columns]
         _timestamp = None
         if timestamp:
             _timestamp = int(timestamp.timestamp() * 1000)
@@ -697,13 +697,13 @@ class HBaseCli:
             AttributeError: 表名格式不合法
         """
         if ns is not None:
-            tableName = TTableName(qualifier=table, ns=ns)
+            tableName = TTableName(qualifier=table.encode(), ns=ns.encode())
         else:
             tabelinfo = table.split(":")
             if len(tabelinfo) == 1:
-                tableName = TTableName(qualifier=table)
+                tableName = TTableName(qualifier=table.encode())
             elif len(tabelinfo) == 2:
-                tableName = TTableName(qualifier=tabelinfo[1], ns=tabelinfo[0])
+                tableName = TTableName(qualifier=tabelinfo[1].encode(), ns=tabelinfo[0].encode())
             else:
                 raise AttributeError(f"parameter table syntax error: {table}")
         sk = None
@@ -738,13 +738,13 @@ class HBaseCli:
             bool: 是否存在
         """
         if ns is not None:
-            tableName = TTableName(qualifier=table, ns=ns)
+            tableName = TTableName(qualifier=table.encode(), ns=ns.encode())
         else:
             tabelinfo = table.split(":")
             if len(tabelinfo) == 1:
-                tableName = TTableName(qualifier=table)
+                tableName = TTableName(qualifier=table.encode())
             elif len(tabelinfo) == 2:
-                tableName = TTableName(qualifier=tabelinfo[1], ns=tabelinfo[0])
+                tableName = TTableName(qualifier=tabelinfo[1].encode(), ns=tabelinfo[0].encode())
             else:
                 raise AttributeError(f"parameter table syntax error: {table}")
         flg = self.client.tableExists(tableName)
@@ -764,13 +764,13 @@ class HBaseCli:
             Dict[str, str]: 描述表基础信息
         """
         if ns is not None:
-            tableName = TTableName(qualifier=table, ns=ns)
+            tableName = TTableName(qualifier=table.encode(), ns=ns.encode())
         else:
             tabelinfo = table.split(":")
             if len(tabelinfo) == 1:
-                tableName = TTableName(qualifier=table)
+                tableName = TTableName(qualifier=table.encode())
             elif len(tabelinfo) == 2:
-                tableName = TTableName(qualifier=tabelinfo[1], ns=tabelinfo[0])
+                tableName = TTableName(qualifier=tabelinfo[1].encode(), ns=tabelinfo[0].encode())
             else:
                 raise AttributeError(f"parameter table syntax error: {table}")
 
@@ -814,13 +814,13 @@ class HBaseCli:
             List[Dict[str, Any]]: 列簇信息
         """
         if ns is not None:
-            tableName = TTableName(qualifier=table, ns=ns)
+            tableName = TTableName(qualifier=table.encode(), ns=ns.encode())
         else:
             tabelinfo = table.split(":")
             if len(tabelinfo) == 1:
-                tableName = TTableName(qualifier=table)
+                tableName = TTableName(qualifier=table.encode())
             elif len(tabelinfo) == 2:
-                tableName = TTableName(qualifier=tabelinfo[1], ns=tabelinfo[0])
+                tableName = TTableName(qualifier=tabelinfo[1].encode(), ns=tabelinfo[0].encode())
             else:
                 raise AttributeError(f"parameter table syntax error: {table}")
 
@@ -896,13 +896,13 @@ class HBaseCli:
             AttributeError: 表名格式不合法
         """
         if ns is not None:
-            tableName = TTableName(qualifier=table, ns=ns)
+            tableName = TTableName(qualifier=table.encode(), ns=ns.encode())
         else:
             tabelinfo = table.split(":")
             if len(tabelinfo) == 1:
-                tableName = TTableName(qualifier=table)
+                tableName = TTableName(qualifier=table.encode())
             elif len(tabelinfo) == 2:
-                tableName = TTableName(qualifier=tabelinfo[1], ns=tabelinfo[0])
+                tableName = TTableName(qualifier=tabelinfo[1].encode(), ns=tabelinfo[0].encode())
             else:
                 raise AttributeError(f"parameter table syntax error: {table}")
         self.client.deleteTable(tableName)
@@ -918,13 +918,13 @@ class HBaseCli:
             AttributeError: 表名格式不合法
         """
         if ns is not None:
-            tableName = TTableName(qualifier=table, ns=ns)
+            tableName = TTableName(qualifier=table.encode(), ns=ns.encode())
         else:
             tabelinfo = table.split(":")
             if len(tabelinfo) == 1:
-                tableName = TTableName(qualifier=table)
+                tableName = TTableName(qualifier=table.encode())
             elif len(tabelinfo) == 2:
-                tableName = TTableName(qualifier=tabelinfo[1], ns=tabelinfo[0])
+                tableName = TTableName(qualifier=tabelinfo[1].encode(), ns=tabelinfo[0].encode())
             else:
                 raise AttributeError(f"parameter table syntax error: {table}")
         self.client.truncateTable(tableName, preserveSplits)
@@ -939,13 +939,13 @@ class HBaseCli:
             AttributeError: 表名格式不合法
         """
         if ns is not None:
-            tableName = TTableName(qualifier=table, ns=ns)
+            tableName = TTableName(qualifier=table.encode(), ns=ns.encode())
         else:
             tabelinfo = table.split(":")
             if len(tabelinfo) == 1:
-                tableName = TTableName(qualifier=table)
+                tableName = TTableName(qualifier=table.encode())
             elif len(tabelinfo) == 2:
-                tableName = TTableName(qualifier=tabelinfo[1], ns=tabelinfo[0])
+                tableName = TTableName(qualifier=tabelinfo[1].encode(), ns=tabelinfo[0].encode())
             else:
                 raise AttributeError(f"parameter table syntax error: {table}")
         self.client.enableTable(tableName)
@@ -960,13 +960,13 @@ class HBaseCli:
             AttributeError: 表名格式不合法
         """
         if ns is not None:
-            tableName = TTableName(qualifier=table, ns=ns)
+            tableName = TTableName(qualifier=table.encode(), ns=ns.encode())
         else:
             tabelinfo = table.split(":")
             if len(tabelinfo) == 1:
-                tableName = TTableName(qualifier=table)
+                tableName = TTableName(qualifier=table.encode())
             elif len(tabelinfo) == 2:
-                tableName = TTableName(qualifier=tabelinfo[1], ns=tabelinfo[0])
+                tableName = TTableName(qualifier=tabelinfo[1].encode(), ns=tabelinfo[0].encode())
             else:
                 raise AttributeError(f"parameter table syntax error: {table}")
         self.client.disableTable(tableName)
@@ -984,13 +984,13 @@ class HBaseCli:
             bool: 是否已激活
         """
         if ns is not None:
-            tableName = TTableName(qualifier=table, ns=ns)
+            tableName = TTableName(qualifier=table.encode(), ns=ns.encode())
         else:
             tabelinfo = table.split(":")
             if len(tabelinfo) == 1:
-                tableName = TTableName(qualifier=table)
+                tableName = TTableName(qualifier=table.encode())
             elif len(tabelinfo) == 2:
-                tableName = TTableName(qualifier=tabelinfo[1], ns=tabelinfo[0])
+                tableName = TTableName(qualifier=tabelinfo[1].encode(), ns=tabelinfo[0].encode())
             else:
                 raise AttributeError(f"parameter table syntax error: {table}")
         return self.client.isTableEnabled(tableName)
@@ -1008,13 +1008,13 @@ class HBaseCli:
             bool: 是否已取消激活
         """
         if ns is not None:
-            tableName = TTableName(qualifier=table, ns=ns)
+            tableName = TTableName(qualifier=table.encode(), ns=ns.encode())
         else:
             tabelinfo = table.split(":")
             if len(tabelinfo) == 1:
-                tableName = TTableName(qualifier=table)
+                tableName = TTableName(qualifier=table.encode())
             elif len(tabelinfo) == 2:
-                tableName = TTableName(qualifier=tabelinfo[1], ns=tabelinfo[0])
+                tableName = TTableName(qualifier=tabelinfo[1].encode(), ns=tabelinfo[0].encode())
             else:
                 raise AttributeError(f"parameter table syntax error: {table}")
         return self.client.isTableDisabled(tableName)
@@ -1032,13 +1032,13 @@ class HBaseCli:
             bool: 是否可用
         """
         if ns is not None:
-            tableName = TTableName(qualifier=table, ns=ns)
+            tableName = TTableName(qualifier=table.encode(), ns=ns.encode())
         else:
             tabelinfo = table.split(":")
             if len(tabelinfo) == 1:
-                tableName = TTableName(qualifier=table)
+                tableName = TTableName(qualifier=table.encode())
             elif len(tabelinfo) == 2:
-                tableName = TTableName(qualifier=tabelinfo[1], ns=tabelinfo[0])
+                tableName = TTableName(qualifier=tabelinfo[1].encode(), ns=tabelinfo[0].encode())
             else:
                 raise AttributeError(f"parameter table syntax error: {table}")
         return self.client.isTableAvailable(tableName)
@@ -1095,13 +1095,13 @@ class HBaseCli:
             AttributeError: 表名格式不合法
         """
         if ns is not None:
-            tableName = TTableName(qualifier=table, ns=ns)
+            tableName = TTableName(qualifier=table.encode(), ns=ns.encode())
         else:
             tabelinfo = table.split(":")
             if len(tabelinfo) == 1:
-                tableName = TTableName(qualifier=table)
+                tableName = TTableName(qualifier=table.encode())
             elif len(tabelinfo) == 2:
-                tableName = TTableName(qualifier=tabelinfo[1], ns=tabelinfo[0])
+                tableName = TTableName(qualifier=tabelinfo[1].encode(), ns=tabelinfo[0].encode())
             else:
                 raise AttributeError(f"parameter table syntax error: {table}")
         _bloomnFilterType = None
@@ -1155,13 +1155,13 @@ class HBaseCli:
             AttributeError: _description_
         """
         if ns is not None:
-            tableName = TTableName(qualifier=table, ns=ns)
+            tableName = TTableName(qualifier=table.encode(), ns=ns.encode())
         else:
             tabelinfo = table.split(":")
             if len(tabelinfo) == 1:
-                tableName = TTableName(qualifier=table)
+                tableName = TTableName(qualifier=table.encode())
             elif len(tabelinfo) == 2:
-                tableName = TTableName(qualifier=tabelinfo[1], ns=tabelinfo[0])
+                tableName = TTableName(qualifier=tabelinfo[1].encode(), ns=tabelinfo[0].encode())
             else:
                 raise AttributeError(f"parameter table syntax error: {table}")
         self.client.deleteColumnFamily(tableName, family.encode("utf-8"))
@@ -1219,13 +1219,13 @@ class HBaseCli:
             AttributeError: 表名格式不合法
         """
         if ns is not None:
-            tableName = TTableName(qualifier=table, ns=ns)
+            tableName = TTableName(qualifier=table.encode(), ns=ns.encode())
         else:
             tabelinfo = table.split(":")
             if len(tabelinfo) == 1:
-                tableName = TTableName(qualifier=table)
+                tableName = TTableName(qualifier=table.encode())
             elif len(tabelinfo) == 2:
-                tableName = TTableName(qualifier=tabelinfo[1], ns=tabelinfo[0])
+                tableName = TTableName(qualifier=tabelinfo[1].encode(), ns=tabelinfo[0].encode())
             else:
                 raise AttributeError(f"parameter table syntax error: {table}")
         _bloomnFilterType = None
@@ -1390,7 +1390,7 @@ class HBaseCli:
         row_bytes = row.encode("utf8")
         _columns = None
         if columns:
-            _columns = [TColumn(family=c.split(":")[0], qualifier=c.split(":")[1]) if len(c.split(":")) == 2 else TColumn(family=c) for c in columns]
+            _columns = [TColumn(family=c.split(":")[0].encode(), qualifier=c.split(":")[1].encode()) if len(c.split(":")) == 2 else TColumn(family=c.encode()) for c in columns]
 
         _timestamp = None
         if timestamp:
@@ -1523,7 +1523,7 @@ class HBaseCli:
                 raise AttributeError(f"parameter table syntax error: {table}")
         _columns = None
         if columns:
-            _columns = [TColumn(family=c.split(":")[0], qualifier=c.split(":")[1]) if len(c.split(":")) == 2 else TColumn(family=c) for c in columns]
+            _columns = [TColumn(family=c.split(":")[0].encode(), qualifier=c.split(":")[1].encode()) if len(c.split(":")) == 2 else TColumn(family=c.encode()) for c in columns]
 
         _timestamp = None
         if timestamp:
@@ -1794,7 +1794,7 @@ class HBaseCli:
         row_bytes = row.encode("utf8")
         _columns = None
         if columns:
-            _columns = [TColumn(family=c.split(":")[0], qualifier=c.split(":")[1]) if len(c.split(":")) == 2 else TColumn(family=c) for c in columns]
+            _columns = [TColumn(family=c.split(":")[0].encode(), qualifier=c.split(":")[1].encode()) if len(c.split(":")) == 2 else TColumn(family=c.encode()) for c in columns]
 
         _timestamp = None
         if timestamp:
@@ -1901,7 +1901,7 @@ class HBaseCli:
         row_bytes = row.encode("utf8")
         _columns = None
         if columns:
-            _columns = [TColumn(family=c.split(":")[0], qualifier=c.split(":")[1]) if len(c.split(":")) == 2 else TColumn(family=c) for c in columns]
+            _columns = [TColumn(family=c.split(":")[0].encode(), qualifier=c.split(":")[1].encode()) if len(c.split(":")) == 2 else TColumn(family=c.encode()) for c in columns]
 
         _timestamp = None
         if timestamp:
@@ -2024,7 +2024,7 @@ class HBaseCli:
         rows_bytes = [row.encode("utf8") for row in rows]
         _columns = None
         if columns:
-            _columns = [TColumn(family=c.split(":")[0], qualifier=c.split(":")[1]) if len(c.split(":")) == 2 else TColumn(family=c) for c in columns]
+            _columns = [TColumn(family=c.split(":")[0].encode(), qualifier=c.split(":")[1].encode()) if len(c.split(":")) == 2 else TColumn(family=c.encode()) for c in columns]
 
         _timestamp = None
         if timestamp:
@@ -2155,7 +2155,7 @@ class HBaseCli:
                 raise AttributeError(f"parameter table syntax error: {table}")
         scan_dict: Dict[str, Any] = {}
         if columns:
-            scan_dict["columns"] = [TColumn(family=c.split(":")[0], qualifier=c.split(":")[1]) if len(c.split(":")) == 2 else TColumn(family=c) for c in columns]
+            scan_dict["columns"] = [TColumn(family=c.split(":")[0].encode(), qualifier=c.split(":")[1].encode()) if len(c.split(":")) == 2 else TColumn(family=c.encode()) for c in columns]
         if caching:
             scan_dict["caching"] = caching
         if maxVersions: