+++ document/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 + 增加了对mutation的支持
 + put操作增加了额外参数
++ 统一了TTableName的输入,现在会全部转换为bytes
++ 统一了TColumn的输入,现在全部会转换为bytes
 
 # 0.0.1